refactor(gsd): extract auto-worktree merge module
